pull out ReplaceRegexMatch tests
